@@ -23,7 +23,7 @@ namespace Iot.Device.SenseHat
         /// Constructs SenseHatJoystick instance
         /// </summary>
         /// <param name="i2cDevice">I2C device used to communicate with the device</param>
-        public SenseHatJoystick(I2cDevice i2cDevice = null)
+        public SenseHatJoystick(I2cDevice? i2cDevice = null)
         {
             _i2c = i2cDevice ?? CreateDefaultI2cDevice();
             Read();
@@ -77,7 +77,7 @@ namespace Iot.Device.SenseHat
         public void Dispose()
         {
             _i2c?.Dispose();
-            _i2c = null;
+            _i2c = null!;
         }
 
         private JoystickState ReadState()
